requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
joblib>=1.3.0
requests>=2.31.0
